to-do app/requirements.txt
Flask
render_template
request
redirect
url_for
session
flash
mysql.connector
wraps
generate_password_hash
check_password_hash